paypal debuggeado y ordenado
